+++ bfabric_app_runner/src/bfabric_app_runner/dispatch/dispatch_single_dataset_flow.py
@@ -21,6 +21,6 @@ class DispatchSingleDatasetFlow:
             logger.error("No dataset found in workunit.")
             return
 
-        dataset = Dataset.find(id=workunit.execution.dataset, client=self._client)
+        dataset = self._client.reader.read_id("dataset", workunit.execution.dataset, expected_type=Dataset)
         path = self.dispatch_job(dataset=dataset, workunit=workunit)
         write_chunks_file(out_dir=self._out_dir, chunks=[path])