+++ bfabric_app_runner/src/bfabric_app_runner/inputs/resolve/_resolve_bfabric_order_fasta_specs.py
@@ -27,7 +27,7 @@ class ResolveBfabricOrderFastaSpecs:
         """Extract FASTA sequence from an order or workunit."""
         # find the order
         if spec.entity == "workunit":
-            workunit = Workunit.find(id=spec.id, client=self._client)
+            workunit = self._client.reader.read_id("workunit", spec.id, expected_type=Workunit)
             if not isinstance(workunit.container, Order):
                 msg = f"Workunit {spec.id} is not associated with an order"
                 if spec.required:
@@ -36,7 +36,7 @@ class ResolveBfabricOrderFastaSpecs:
                 return ""
             order = workunit.container
         elif spec.entity == "order":
-            order = Order.find(id=spec.id, client=self._client)
+            order = self._client.reader.read_id("order", spec.id, expected_type=Order)
             if order is None:
                 msg = f"Order {spec.id} not found"
                 if spec.required: